+++ kipper/core/src/compiler/index.ts
@@ -1,0 +1,21 @@
+/**
+ * The main module for the Kipper compiler
+ * @author Luna Klatzer
+ * @copyright 2021-2022 Luna Klatzer
+ * @since 0.0.1
+ */
+
+export * as tokens from "./tokens";
+export * as logic from "./logic";
+export * as target from "./target";
+export * as parser from "./parser";
+export * from "./tokens";
+export * from "./logic";
+export * from "./antlr-error-listener";
+export * from "./parse-stream";
+export * from "./listener";
+export * from "./program-ctx";
+export * from "./compiler";
+export * from "./semantic-analyser";
+export * from "./code-generator";
+export * from "./const";